test: convert to the standard Ruby layout for tests
