+++ Rakefile
@@ -5,9 +5,8 @@ task :default
 
 Rake::TestTask.new(:test) do |t|
     t.libs << "lib"
-    t.libs << "."
-    t.ruby_opts << '-w'
-    t.test_files = FileList['test/**/test_*.rb']
+    t.libs << "test"
+    t.test_files = FileList['test/**/*_test.rb']
 end
 
 task :gem => :build
